+++ packages/ethereum-contracts/contracts/interfaces/superfluid/ISuperToken.sol
@@ -15,6 +15,12 @@ import { IPoolMemberNFT } from "../agreements/gdav1/IPoolMemberNFT.sol";
  */
 interface ISuperToken is ISuperfluidToken, IERC20Metadata, IERC777 {
 
+    struct AdminOverride {
+        address admin;
+        /// @note we use a struct so the 12 remaining
+        /// packed bytes may be used later on
+    }
+
     /**************************************************************************
      * Errors
      *************************************************************************/
@@ -23,7 +29,7 @@ interface ISuperToken is ISuperfluidToken, IERC20Metadata, IERC777 {
     error SUPER_TOKEN_INFLATIONARY_DEFLATIONARY_NOT_SUPPORTED(); // 0xe3e13698
     error SUPER_TOKEN_NO_UNDERLYING_TOKEN();                     // 0xf79cf656
     error SUPER_TOKEN_ONLY_SELF();                               // 0x7ffa6648
-    error SUPER_TOKEN_ONLY_HOST();                               // 0x98f73704
+    error SUPER_TOKEN_ONLY_ADMIN();                              // 0x0484acab
     error SUPER_TOKEN_ONLY_GOV_OWNER();                          // 0xd9c7ed08
     error SUPER_TOKEN_APPROVE_FROM_ZERO_ADDRESS();               // 0x81638627
     error SUPER_TOKEN_APPROVE_TO_ZERO_ADDRESS();                 // 0xdf070274
@@ -42,6 +48,32 @@ interface ISuperToken is ISuperfluidToken, IERC20Metadata, IERC777 {
         string calldata n,
         string calldata s
     ) external;
+
+    /**
+     * @dev Initialize the contract with admin override
+     */
+    function initializeWithAdminOverride(
+        IERC20 underlyingToken,
+        uint8 underlyingDecimals,
+        string calldata n,
+        string calldata s,
+        address adminOverride
+    ) external;
+
+    /**
+     * @notice Changes the admin override for the SuperToken
+     * @dev Only the current admin can call this function
+     * if adminOverride is address(0), it is implicitly the host address
+     * @param newAdmin New admin override address
+     */
+    function changeAdmin(address newAdmin) external;
+
+    event AdminChanged(address indexed oldAdmin, address indexed newAdmin);
+
+    /**
+     * @dev Returns the admin override struct for the SuperToken
+     */
+    function getAdminOverride() external view returns (AdminOverride memory);
 
     /**************************************************************************
     * Immutable variables