+++ packages/ethereum-contracts/dev-scripts/deploy-contracts-and-token.js
@@ -8,7 +8,7 @@ const {
 async function deployContractsAndToken() {
     const [Deployer] = await ethers.getSigners();
 
-    const {frameworkDeployer: deployer } = await deployTestFramework();
+    const {frameworkDeployer: deployer} = await deployTestFramework();
     const framework = await deployer.getFramework();
 
     const resolver = await ethers.getContractAt(
@@ -18,7 +18,7 @@ async function deployContractsAndToken() {
 
     await deployer
         .connect(Deployer)
-        .deployWrapperSuperToken(
+        ["deployWrapperSuperToken(string,string,uint8,uint256)"](
             "Fake DAI",
             "fDAI",
             18,